// src/app/app.interface.ts
export interface StudentData {
    key: string;
    tenthmarks: number;
    batch: number;
    branch: string;
    btechMarks: number;
    collegeName: string;
    email: string;
    gender: string;
    interMarks: number;
    name: string;
    phoneNo: number;
    placementReady: string;
    status: string;
    yearOfPassout: number;
}